Add tab_controller.py. And revised persephonep.py and func_persephonep.py. Changes are very simple. 1) separate classes which control TabWidget from persephonep.py into tab_controller.py. 2) some global parameters is or will be called via functions.
